Fixed bug in tests for ratio functions
The bug occurred in tests for FunctionOnRatioOfRuns and FunctionOnRatioOfInversions.
This was due to the way the apply method of the DoubleToIntFunction proxy instance worked:
The apply method of the two aforementioned classes calculate the ratio of the number of runs in a sequence to the
length of said sequence and the ratio of inversions in a sequence to the number of possible inversions, respectively.
The results should then be passed to the apply method of the DoubleToIntFunction proxy. However, since these values
would almost exclusively be less than 1 and the proxy would simply return the value as an int, the tests checked
if 0 == 0 was true most of the time.

The proxy apply method now multiplies the given value with 10^8 and casts it to int, which should be enough precision.